Set ID for Log file setting wrapper.
We missed standard ID used for setting's form fields. It's used
in the code which updates visibility of setting which depends
on state of another setting.

Missing ID breaks JS code in Magento prior 1.9 release so it
prevents from initialising other parts of form and other
functionality such as saving config